Refactor: BottomBar 수정
